Removing comments, couts and renaming the TimeCalibration_dev to TimeCalibration
